requirements.txt
torch
torchvision
streamlit
opencv-python
scikit-learn
Pillow
faiss-cpu
numpy
